Adding watch goal to trigger a compile when there are changes
